requirements.txt
boto3==1.17.98
PyYAML==5.4.1

